Fixed Issue #99. Made the menteelogin.html page editable in the Admin page, specifically for the heaing, description and image, not the FB/Twitter feed. User can now add a single post in Content_Summarys that will display the menteelogin.html . Images will delete if changed from original.

+++ project/fifty_fifty/content/models.py
@@ -17,21 +17,29 @@ class Post(models.Model):
 
 class Mentor(models.Model):
     title = models.CharField(max_length=200)
+    description = models.TextField(null=True)
+    image = models.ImageField(upload_to='mentor/images/', default="blog_images/stock_photo.jpg")
     docfile = models.FileField(upload_to='mentor/', null=True)
 
     def __str__(self):
         return self.title
 
+
+###DON'T CHANGE THIS MODEL
 class Mentee(models.Model):
     title = models.CharField(max_length=200)
+    description = models.TextField(null=True)
+    image = models.ImageField(upload_to='mentee/images/', default="blog_images/stock_photo.jpg")
     docfile = models.FileField(upload_to='mentee/', null=True)
 
     def __str__(self):
         return self.title
 
-class Training(models.Model):
+class Content_Summary(models.Model):
     title = models.CharField(max_length=200)
-    docfile = models.FileField(upload_to='training/', null=True)
+    description = models.TextField(null=True)
+    image = models.ImageField(upload_to='content_summary/images/', default="blog_images/stock_photo.jpg")
+    ###docfile = models.FileField(upload_to='content_summary/', null=True)
 
     def __str__(self):
         return self.title
@@ -41,16 +49,24 @@ class Training(models.Model):
 @receiver(post_delete, sender=Post)
 @receiver(post_delete, sender=Mentor)
 @receiver(post_delete, sender=Mentee)
-@receiver(post_delete, sender=Training)
+@receiver(post_delete, sender=Content_Summary)
 def delete(sender, instance, **kwargs):
     # Pass false so FileField doesn't save the model.
     instance.docfile.delete(False)
 
-#Edit content from folder
+#Delete images from folder
+@receiver(post_delete, sender=Post)
+@receiver(post_delete, sender=Mentor)
+@receiver(post_delete, sender=Mentee)
+@receiver(post_delete, sender=Content_Summary)
+def delete(sender, instance, **kwargs):
+    # Pass false so FileField doesn't save the model.
+    instance.image.delete(False)
+
+#Edit docfile from folder
 @receiver(pre_save, sender=Post)
 @receiver(pre_save, sender=Mentor)
 @receiver(pre_save, sender=Mentee)
-@receiver(pre_save, sender=Training)
 def auto_delete_file_on_change(sender, instance, **kwargs):
     if not instance.pk:
         return False
@@ -59,6 +75,23 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
     except sender.DoesNotExist:
         return False
     new_file = instance.docfile
+    if not old_file == new_file:
+        if os.path.isfile(old_file.path):
+            os.remove(old_file.path)
+
+#Edit images from folder
+@receiver(pre_save, sender=Post)
+@receiver(pre_save, sender=Mentor)
+@receiver(pre_save, sender=Mentee)
+@receiver(pre_save, sender=Content_Summary)
+def auto_delete_file_on_change(sender, instance, **kwargs):
+    if not instance.pk:
+        return False
+    try:
+        old_file = sender.objects.get(pk=instance.pk).image
+    except sender.DoesNotExist:
+        return False
+    new_file = instance.image
     if not old_file == new_file:
         if os.path.isfile(old_file.path):
             os.remove(old_file.path)
